requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
argcomplete==3.6.2
attrs==25.3.0
boto3==1.40.21
botocore==1.40.21
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cohere==5.17.0
colorama==0.4.6
colorlog==6.9.0
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.116.1
fastavro==1.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
genai-prices==0.0.24
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-genai==1.32.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.13.0
groq==0.31.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
logfire-api==4.3.6
markdown-it-py==4.0.0
mcp==1.13.1
mdurl==0.1.2
mistralai==1.9.9
multidict==6.6.4
mysql-connector==2.2.9
nexus-rpc==1.1.0
openai==1.102.0
opentelemetry-api==1.36.0
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.8.1
pydantic-ai-slim==0.8.1
pydantic-evals==0.8.1
pydantic-graph==0.8.1
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
temporalio==1.16.0
tenacity==9.1.2
tokenizers==0.22.0
tqdm==4.67.1
types-protobuf==6.30.2.20250822
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
